requirements.txt
aiofiles==24.1.0
html5tagger==1.3.0
httptools==0.6.4
multidict==6.1.0
python-dotenv==1.0.1
sanic==24.6.0
sanic-routing==23.12.0
setuptools==75.6.0
tracerite==1.1.1
typing_extensions==4.12.2
ujson==5.10.0
uvloop==0.21.0
websockets==14.1
